+++ WarcraftGuild/ExternalApi/BlizzardApi/Json/Common/TypeJson.cs
@@ -5,7 +5,7 @@ namespace WarcraftGuild.BlizzardApi.Json
     public class TypeJson : BlizzardApiJsonResponse
     {
         [JsonPropertyName("id")]
-        public ulong Id { get; set; }
+        public uint Id { get; set; }
 
         [JsonPropertyName("type")]
         public string Type { get; set; }
